Make VFS copy faster, by doing concurrent copies
During copies, especially with small files, a significant portion of
the time is spent doing metadata ops. With reflinks, these metadata
ops become the vast majority of the time, as opposed to a negligible
amount of time.

Locking of the metadata (Extents) occurs on a per-inode basis, and
also on a dentry basis, for this reason, we send workloads to different
goroutines based on the directory in order to avoid potential contention.

Signed-off-by: Sargun Dhillon <sargun@sargun.me>